Unification of repositories in a single repository
Initially, the objective was to have this repository as host for others "sub-repositories". Each project developed in the discipline would have its own repository. 

Now the paradigm has changed and this repository will host all that's developed with the organization and management of each project being made in the branches.